Create workspace for common functionality. More refactoring.
