+++ carina/package.json
@@ -1,14 +1,12 @@
 {
-  "name": "carina",
+  "name": "@minids/carina",
   "version": "0.0.0",
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "^6.2.1",
     "@fortawesome/free-brands-svg-icons": "^6.2.1",
     "@fortawesome/free-solid-svg-icons": "^6.2.1",
     "@fortawesome/vue-fontawesome": "^3",
-    "@wwtelescope/engine-pinia": "^0.1.0",
-    "screenfull": "^6.0.2",
-    "vue": "^3",
+    "@minids/common": "workspace:0.0.0",
     "vuetify": "^3.1.1",
     "webpack-plugin-vuetify": "^2.0.0"
   },
