+++ carina/tsconfig.json
@@ -1,6 +1,17 @@
 {
-  "extends": "../tsconfig.json",
   "compilerOptions": {
+    "target": "es5",
+    "module": "esnext",
+    "composite": true,
+    "strict": true,
+    "jsx": "preserve",
+    "importHelpers": true,
+    "downlevelIteration": true,
+    "moduleResolution": "node",
+    "experimentalDecorators": true,
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
+    "sourceMap": true,
     "baseUrl": ".",
     "outDir": "dist",
     "paths": {
@@ -15,9 +26,6 @@
     ]
   },
   "include": [
-    "../components/*.ts",
-    "../components/*.vue",
-    "../utils/*.ts",
     "plugins/*.ts",
     "src/**/*.ts",
     "src/**/*.vue"
