+++ common/src/index.ts
@@ -1,0 +1,7 @@
+import { KeyboardControlSettings } from "./keyboard";
+import MiniDSBase from "./components/MiniDSBase";
+
+export {
+  KeyboardControlSettings,
+  MiniDSBase
+};
