+++ common/webpack.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  module: {
+    rules: [
+      {
+        test: "/\.(png|jpg)$/i",
+        type: 'asset/resources'
+      }
+    ]
+  }
+}
